Refactor: package classes
Packaged classes according to handler, model, util for authentication and services for AuthClient, CatalogueClient related classes
