--- src/main/java/iudx/data/marketplace/authenticator/authorization/Method.java
@@ -1,8 +1,0 @@
-package iudx.data.marketplace.authenticator.authorization;
-
-public enum Method {
-  GET,
-  POST,
-  PUT,
-  DELETE
-}
