+++ src/main/java/iudx/data/marketplace/authenticator/handler/AuthorizationRequest.java
@@ -1,4 +1,4 @@
-package iudx.data.marketplace.authenticator.authorization;
+package iudx.data.marketplace.authenticator.handler;
 
 import java.util.Objects;
