+++ src/main/java/iudx/data/marketplace/authenticator/handler/Method.java
@@ -1,0 +1,8 @@
+package iudx.data.marketplace.authenticator.handler;
+
+public enum Method {
+  GET,
+  POST,
+  PUT,
+  DELETE
+}
