+++ src/test/java/iudx/data/marketplace/authenticator/TestAuthClient.java
@@ -1,6 +1,7 @@
 package iudx.data.marketplace.authenticator;
 
 import io.vertx.core.buffer.Buffer;
+import iudx.data.marketplace.aaaService.AuthClient;
 import iudx.data.marketplace.apiserver.handlers.AuthHandler;
 import iudx.data.marketplace.apiserver.util.Role;
 import iudx.data.marketplace.common.Api;
@@ -27,15 +28,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.stubbing.Answer;
-import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.PostgreSQLContainer;
 
 import static iudx.data.marketplace.apiserver.util.Constants.ROLE;
-import static iudx.data.marketplace.product.util.Constants.RESULTS;
 import static iudx.data.marketplace.product.util.Constants.TYPE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
