+++ src/test/java/iudx/data/marketplace/razorpay/TestRazorpayServiceImpl.java
@@ -7,16 +7,9 @@ import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
-import iudx.data.marketplace.Util;
-import iudx.data.marketplace.apiserver.exceptions.DxRuntimeException;
-import iudx.data.marketplace.authenticator.authorization.ConsumerAuthStatergy;
-import iudx.data.marketplace.authenticator.authorization.IudxRole;
-import iudx.data.marketplace.authenticator.authorization.ProviderAuthStatergy;
 import iudx.data.marketplace.common.Api;
 import iudx.data.marketplace.common.ResponseUrn;
 import iudx.data.marketplace.policies.User;
-import iudx.data.marketplace.policies.VerifyPolicy;
-import iudx.data.marketplace.policy.TestCreatePolicy;
 import iudx.data.marketplace.postgres.PostgresService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -24,7 +17,6 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
